Update debugger and compiler paths, fix type casts
Changed debugger and compiler paths in VSCode configs to use msys64 instead of gcc. Fixed type casts in dump_misc.c for regex_search calls. Updated IMAGE_SYM_CLASS_END_OF_FUNCTION enum value to (BYTE)-1 in pe_flags.h. Swapped test subject path in main.c for kernel32.dll.

+++ src/include/pe_flags.h
@@ -161,7 +161,7 @@ typedef enum _IMAGE_SYMBOL_DERIVED_TYPE {
 // ---------------------------------------------------------------------
 #ifndef IMAGE_SYM_CLASS_END_OF_FUNCTION
     typedef enum _IMAGE_SYMBOL_STORAGE_CLASS {
-        IMAGE_SYM_CLASS_END_OF_FUNCTION      = -1,
+        IMAGE_SYM_CLASS_END_OF_FUNCTION      = (BYTE)-1,
         IMAGE_SYM_CLASS_NULL                 = 0x00,
         IMAGE_SYM_CLASS_AUTOMATIC            = 0x01,
         IMAGE_SYM_CLASS_EXTERNAL             = 0x02,
